requirements.txt
openai
langchain>=0.1.16
langchain-community>=0.0.34
langchain-openai>=0.1.6
langchain-huggingface>=0.0.6
faiss-cpu
pandas
python-dotenv
transformers
sentence-transformers
